refactor: refactorized quoteProject.tsx in components per phases
